requirements.txt
fastapi
uvicorn

# Authentication dependencies
python-jose[cryptography]
passlib[bcrypt]
